fix: move long AI requests outside service worker and close review gaps
