+++ .github/workflows/release.yml
@@ -27,6 +27,7 @@ jobs:
           ZIP_NAME="resume-pro-${GITHUB_REF_NAME}.zip"
           git archive --format=zip --output="$ZIP_NAME" HEAD \
             manifest.json background.js content.js content.css ai-helpers.js form-agent.js \
+            ai-worker.js ai-host.js ai-host.html ai-client.js \
             resume-utils.js popup.html popup.css popup.js xlsx.full.min.js \
             mammoth.browser.min.js vendor icons README.md LICENSE
           echo "ZIP_NAME=$ZIP_NAME" >> $GITHUB_ENV
