+++ README.md
@@ -423,6 +423,7 @@ API Key 请勿分享给他人，也不要提交到公开 GitHub Issue。
 点击「一键 AI 填写」后，按钮显示扫描、AI 匹配秒数和填写阶段。完成或失败后，展开的「填写诊断」保留扫描、匹配往返、API（含响应体读取）、填写和总耗时，以及字段数量和用户 prompt 字节数；可在只读文本框中选择复制。未执行或未取得的阶段不会显示成零耗时。
 
 - AI 等待超过 90 秒仅显示慢响应提醒，不自动取消。用户可随时点击「取消 AI 等待（保留本地匹配）」主动终止请求，包括响应体读取；取消不保证上游停止计算或停止计费。浏览器扩展生命周期、网络或上游服务仍可能自行中断请求。
+- v0.3.0 需要 Chrome / Edge 116 或更新版本，并使用 `offscreen` 权限。AI 网络请求由扩展隐藏文档中的独立 Worker 执行，不放在 Service Worker 中等待响应头；简历解析也使用同一请求进程。浏览器退出、扩展重载仍会中断任务。
 - HTTP、网络、用户取消、返回格式错误分别提示。AI 失败或手动取消时仍使用已验证的本地匹配结果，明确显示部分完成；不会自动重试或更改用户接口配置。慢响应提示说明可能与上游模型、中转、网络和输入量有关，不在缺少实测证据时断言一定是上游故障。
 - 对明确的教育、工作、项目等字段，只发送相关的完整简历分组，保留多段经历及其关联字段。含糊、跨领域或无相关候选时回退全量，自定义未知分组始终保留。不采用固定 Top N 截断。
 - 诊断摘要只包含版本、计数、字节数、耗时和标准错误类别，不包含接口地址、API Key、简历内容或上游错误原文。截图时仍需注意周围网页和模板本身的个人信息。
@@ -447,7 +448,10 @@ API Key 请勿分享给他人，也不要提交到公开 GitHub Issue。
 manifest.json          插件清单
 popup.html/css/js      模板管理、AI 配置、简历解析界面
 content.js/css         页面注入、字段识别、侧边栏和填写逻辑
-background.js          AI 接口调用和消息路由
+background.js          扩展动作和隐藏请求进程启动
+ai-host.html/js        隐藏文档及消息转发
+ai-worker.js           AI 匹配、解析和可取消网络请求
+ai-client.js           请求进程准备及客户端取消顺序控制
 ai-helpers.js          字段匹配、数据清洗和辅助规则
 form-agent.js          受限新增计划、分组识别和执行检查
 resume-utils.js        PDF 文本、版本比较和错误提示辅助逻辑
