+++ ai-host.html
@@ -1,0 +1,4 @@
+<!doctype html>
+<meta charset="utf-8">
+<title>Resume Pro AI request host</title>
+<script src="ai-host.js"></script>
