+++ docs/ai-repeat-validation.md
@@ -29,3 +29,11 @@ Both scenarios used real confirm previews and asynchronous insertion of fieldset
 ## Deliberate limits
 
 Opt-in preview, at most five additions, exact supported section/button structure, visible text inputs/textareas only. No arbitrary scripts, navigation, submit, deletion or automatic rollback. A short value readback does not prove server-side persistence. Native/browser lifecycle termination remains possible. Real websites using custom controls, modals, iframes, shadow DOM or reactive row replacement require separately verified adapters; users should use the manual fallback and review before submitting.
+
+## Release review follow-up (2026-09-06)
+
+The request path now runs in a dedicated worker hosted by an extension offscreen document. The service worker only ensures the host exists; content/popup requests and responses do not wait on a service-worker fetch. Chrome/Edge 116+ and the offscreen permission are required.
+
+An actual unpacked extension was loaded into an isolated Edge profile (no Chrome API shims). A localhost server deliberately withheld HTTP response headers for 100 seconds. The extension completed successfully in 100.057 seconds, with API time 100.016 seconds and the expected match. A second request was manually cancelled; the server connection closed, the error category was `cancelled`, and the local name match was retained. No external AI credits were consumed for these lifecycle tests.
+
+107 automated tests pass after review fixes: authorization/current-employer candidate retention, graduate-school classification, suffix-numbered records, same-anchor uniqueness suffixes, complete target counts, offscreen startup concurrency, message correlation, startup/cancellation ordering and dependency packaging. All identified review findings have regression coverage.
