+++ popup.html
@@ -129,6 +129,7 @@
     <script src="mammoth.browser.min.js"></script>
     <script src="ai-helpers.js"></script>
     <script src="resume-utils.js"></script>
+    <script src="ai-client.js"></script>
     <script src="popup.js"></script>
   </body>
 </html>
