+++ tests/issue-8.test.js
@@ -175,7 +175,7 @@ test("bundled PDF.js rejects invalid PDF bytes", async () => {
 
 test("runtime source sends extracted text and contains no PDF-as-image path", () => {
   const popupSource = fs.readFileSync(path.join(__dirname, "..", "popup.js"), "utf8");
-  const backgroundSource = fs.readFileSync(path.join(__dirname, "..", "background.js"), "utf8");
+  const backgroundSource = fs.readFileSync(path.join(__dirname, "..", "ai-worker.js"), "utf8");
   const contentSource = fs.readFileSync(path.join(__dirname, "..", "content.js"), "utf8");
   const htmlSource = fs.readFileSync(path.join(__dirname, "..", "popup.html"), "utf8");
